pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "pwdlib[argon2]>=0.3.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
    "ty>=0.0.9",
]

[tool.ruff]
exclude = ["alembic/versions"]

[tool.ruff.lint]
select = [
    "E", "F",     # Default Error & Pyflakes
    "RUF",        # Ruff-specific rules
    "W",          # pycodestyle warnings
    "C4",         # flake8-comprehensions
    "UP",         # pyupgrade
    "ARG001",     # unused arguments in functions
    "ANN",        # flake8-annotations (Strict type hinting)
    "UP",         # pyupgrade (Newest type syntax: e.g., int | None)
    "B",          # flake8-bugbear (Common logic errors)
    "I",          # isort (Import sorting)
    "FAST",       # FastAPI,
    "ASYNC",      # flake8-async (crucial for FastAPI/Async SQLAlchemy)
]
